% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_comments.R
\name{collect_comments}
\alias{collect_comments}
\title{Collect the body and comments from a Github issue}
\usage{
collect_comments(url = "https://github.com/JiaxiangBU/add2gh/issues/2",
  repo_owner = "JiaxiangBU",
  repo_name = basename(usethis::proj_path()), out_file_path = c(FALSE,
  here::here("output/tmp.Rmd")), set_number = 1)
}
\arguments{
\item{url}{The Github issue url.}

\item{repo_owner}{The Github id of this repository.}

\item{repo_name}{The Github repository name.}

\item{out_file_path}{The saved path.}

\item{set_number}{The Github issue number.}
}
\value{
The comments from a specific issue.
}
\description{
Collect the body and comments from a Github issue
}
\examples{
\donttest{collect_comments("https://github.com/JiaxiangBU/add2gh/issues/2")}
}
