% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_github_repo.R
\name{create_github_repo}
\alias{create_github_repo}
\title{Create a GitHub repository by one function.}
\usage{
create_github_repo(repo_name, is_project, private_token = private_token,
  username = "JiaxiangBU")
}
\description{
Create a GitHub repository by one function.
}
